refactor(spec): split validate-structure into layers, scales, and facet
Separate data-free grammar checks by concern (layer geometry, color scheme
family rules, facet wrap/grid form) behind a thin validate-structure barrel
so validate() import paths stay stable.
