+++ .changeset/spec-validate-structure-split.md
@@ -1,0 +1,7 @@
+---
+"@ggsvelte/spec": patch
+---
+
+<!-- markdownlint-disable MD041 -->
+
+Split data-free structural validation into layer, color-scheme, and facet modules behind a stable validate-structure barrel. Public validate() behavior is unchanged.
